+++ java/com/google/re2j/Prog.java
@@ -7,30 +7,43 @@
 
 package com.google.re2j;
 
+import java.util.Arrays;
+
 /**
  * A Prog is a compiled regular expression program.
  */
 final class Prog {
 
-  
-  final Inst[] inst;
-  final int start; // index of start instruction
-  final int numCap; // number of CAPTURE insts in re
+  Inst[] inst = new Inst[10];
+  int instSize = 0;
+  int start; // index of start instruction
+  int numCap = 2; // number of CAPTURE insts in re
                   // 2 => implicit ( and ) for whole match $0
 
   // Constructs an empty program.
-  Prog(Inst[] inst, int start, int numCap) {
-    this.inst = inst;
-    this.start = start;
-    this.numCap = numCap;
+  Prog() {}
+
+  // Returns the instruction at the specified pc.
+  // Precondition: pc > 0 && pc < numInst().
+  Inst getInst(int pc) {
+    return inst[pc];
   }
 
   // Returns the number of instructions in this program.
   int numInst() {
-    return inst.length;
+    return instSize;
   }
 
-  
+  // Adds a new instruction to this program, with operator |op| and |pc| equal
+  // to |numInst()|.
+  void addInst(int op) {
+    if (instSize >= inst.length) {
+      inst = Arrays.copyOf(inst, inst.length * 2); 
+    }
+    inst[instSize] = new Inst(op);
+    instSize++;
+  }
+
   // skipNop() follows any no-op or capturing instructions and returns the
   // resulting instruction.
   Inst skipNop(int pc) {
@@ -88,11 +101,70 @@ final class Prog {
     return flag;
   }
 
+  // --- Patch list ---
+
+  // A patchlist is a list of instruction pointers that need to be filled in
+  // (patched).  Because the pointers haven't been filled in yet, we can reuse
+  // their storage to hold the list.  It's kind of sleazy, but works well in
+  // practice.  See http://swtch.com/~rsc/regexp/regexp1.html for inspiration.
+
+  // These aren't really pointers: they're integers, so we can reinterpret them
+  // this way without using package unsafe.  A value l denotes p.inst[l>>1].out
+  // (l&1==0) or .arg (l&1==1).  l == 0 denotes the empty list, okay because we
+  // start every program with a fail instruction, so we'll never want to point
+  // at its output link.
+
+  int next(int l) {
+    Inst i = inst[l >> 1];
+    if ((l & 1) == 0) {
+      return i.out;
+    }
+    return i.arg;
+  }
+
+  void patch(int l, int val) {
+    while (l != 0) {
+      Inst i = inst[l >> 1];
+      if ((l & 1) == 0) {
+        l = i.out;
+        i.out = val;
+      } else {
+        l = i.arg;
+        i.arg = val;
+      }
+    }
+  }
+
+  int append(int l1, int l2) {
+    if (l1 == 0) {
+      return l2;
+    }
+    if (l2 == 0) {
+      return l1;
+    }
+    int last = l1;
+    for (;;) {
+      int next = next(last);
+      if (next == 0) {
+        break;
+      }
+      last = next;
+    }
+    Inst i = inst[last>>1];
+    if ((last & 1) == 0) {
+      i.out = l2;
+    } else {
+      i.arg = l2;
+    }
+    return l1;
+  }
+
+  // ---
 
   @Override
   public String toString() {
     StringBuilder out = new StringBuilder();
-    for (int pc = 0; pc < inst.length; ++pc) {
+    for (int pc = 0; pc < instSize; ++pc) {
       int len = out.length();
       out.append(pc);
       if (pc == start) {
